demo/requirements.txt
fpdf2==2.8.2
pdfplumber==0.11.4
